+++ project_static/mango_app/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path
+from django.urls import path, re_path
 from . import views
 
 app_name = 'mango_app'  # Added app_name for URL namespacing
@@ -10,4 +10,7 @@ urlpatterns = [
     path('diseases/', views.DiseasesView.as_view(), name='diseases'),
     path('surveillance/', views.SurveillanceView.as_view(), name='surveillance'),
     path('about/', views.AboutView.as_view(), name='about'),
+    path('projects/', views.ProjectListView.as_view(), name='projects'),
+    # URL pattern with regex for project detail pages
+    re_path(r'^projects/(?P<item_id>\d+)/$', views.MangoItemDetailView.as_view(), name='project_detail'),
 ]